+++ src/main/kotlin/me/branchpanic/mods/stockpile/Stockpile.kt
@@ -1,5 +1,6 @@
 package me.branchpanic.mods.stockpile
 
+import me.branchpanic.mods.stockpile.api.upgrade.UpgradeRegistry
 import me.branchpanic.mods.stockpile.api.upgrade.UpgradeType
 import me.branchpanic.mods.stockpile.content.block.AttackableBlockCallback
 import me.branchpanic.mods.stockpile.content.block.ItemBarrelBlock
@@ -13,7 +14,6 @@ import me.branchpanic.mods.stockpile.content.upgrade.CapacityUpgrade
 import me.branchpanic.mods.stockpile.content.upgrade.MultiplierUpgrade
 import me.branchpanic.mods.stockpile.content.upgrade.TrashUpgrade
 import me.branchpanic.mods.stockpile.content.upgrade.UpgradeInstallerCallback
-import me.branchpanic.mods.stockpile.impl.upgrade.UpgradeRegistry
 import net.fabricmc.api.ModInitializer
 import net.fabricmc.fabric.api.client.itemgroup.FabricItemGroupBuilder
 import net.fabricmc.fabric.api.event.player.AttackBlockCallback
@@ -71,7 +71,7 @@ object Stockpile : ModInitializer {
             Registry.register(Registry.BLOCK, id, block)
 
             if (ITEMS.keys.none { itemId -> itemId == id }) {
-                Registry.register(Registry.ITEM, id, BlockItem(block, ITEM_SETTINGS))
+                Registry.register(Registry.ITEM, id, BlockItem(block, ITEM_SETTINGS.maxDamage(0).maxCount(8)))
             }
         }