+++ src/main/kotlin/me/branchpanic/mods/stockpile/content/blockentity/ItemBarrelBlockEntity.kt
@@ -1,19 +1,19 @@
 package me.branchpanic.mods.stockpile.content.blockentity
 
-import me.branchpanic.mods.stockpile.Stockpile
+import me.branchpanic.mods.stockpile.api.AbstractBarrelBlockEntity
+import me.branchpanic.mods.stockpile.api.BarrelTransactionAmount
 import me.branchpanic.mods.stockpile.api.upgrade.Upgrade
 import me.branchpanic.mods.stockpile.api.upgrade.UpgradeContainer
+import me.branchpanic.mods.stockpile.api.upgrade.UpgradeRegistry
 import me.branchpanic.mods.stockpile.api.upgrade.barrel.ItemBarrelUpgrade
 import me.branchpanic.mods.stockpile.content.block.ItemBarrelBlock
 import me.branchpanic.mods.stockpile.content.upgrade.TrashUpgrade
-import me.branchpanic.mods.stockpile.giveTo
+import me.branchpanic.mods.stockpile.extension.giveTo
 import me.branchpanic.mods.stockpile.impl.attribute.FixedMassItemInv
 import me.branchpanic.mods.stockpile.impl.attribute.UnrestrictedInventoryFixedWrapper
-import me.branchpanic.mods.stockpile.impl.storage.MassItemStorage
-import me.branchpanic.mods.stockpile.impl.upgrade.UpgradeRegistry
+import me.branchpanic.mods.stockpile.impl.storage.*
 import net.fabricmc.fabric.api.block.entity.BlockEntityClientSerializable
 import net.fabricmc.fabric.api.util.NbtType
-import net.minecraft.block.entity.BlockEntity
 import net.minecraft.block.entity.BlockEntityType
 import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.inventory.Inventory
@@ -22,252 +22,216 @@ import net.minecraft.nbt.CompoundTag
 import net.minecraft.nbt.ListTag
 import net.minecraft.sound.SoundCategory
 import net.minecraft.sound.SoundEvents
-import net.minecraft.text.Text
 import net.minecraft.text.TranslatableText
 import net.minecraft.util.Hand
-import java.text.NumberFormat
-import java.util.*
+import net.minecraft.util.Identifier
+import net.minecraft.util.registry.Registry
 import java.util.function.Supplier
+import kotlin.math.max
 import kotlin.math.min
 
 class ItemBarrelBlockEntity(
-    private var storage: MassItemStorage = MassItemStorage(
-        DEFAULT_CAPACITY_STACKS
-    ),
-    override var appliedUpgrades: List<ItemBarrelUpgrade> = emptyList()
-) :
-    BlockEntity(TYPE),
-    BlockEntityClientSerializable,
-    UpgradeContainer,
-    Inventory {
-    val invAttribute: FixedMassItemInv = FixedMassItemInv(storage)
-    private val invWrapper = UnrestrictedInventoryFixedWrapper(invAttribute)
+    internal var clearWhenEmpty: Boolean = true,
+    override var appliedUpgrades: List<Upgrade> = emptyList(),
+    override var maxUpgrades: Int = DEFAULT_MAX_UPGRADES
+) : AbstractBarrelBlockEntity<ItemStack>(
+    storage = MassItemStackStorage(ItemStackQuantizer.NONE, DEFAULT_CAPACITY_STACKS),
+    doubleClickThresholdMs = 1000,
+    type = TYPE
+), BlockEntityClientSerializable, Inventory, UpgradeContainer {
 
-    init {
-        invAttribute.addListener({ _, _, _, _ -> markDirty() }, {})
+    override fun isUpgradeTypeAllowed(u: Upgrade): Boolean = u is ItemBarrelUpgrade
+
+    override fun pushUpgrade(u: Upgrade) {
+        appliedUpgrades += u
+        markDirty()
     }
 
-    override fun getInvStack(slot: Int): ItemStack = invWrapper.getInvStack(slot)
-
-    override fun clear() = invWrapper.clear()
-
-    override fun setInvStack(p0: Int, p1: ItemStack?) = invWrapper.setInvStack(p0, p1)
-
-    override fun removeInvStack(p0: Int): ItemStack = invWrapper.removeInvStack(p0)
-
-    override fun canPlayerUseInv(p0: PlayerEntity?): Boolean = invWrapper.canPlayerUseInv(p0)
-
-    override fun getInvSize(): Int = invWrapper.invSize
-
-    override fun takeInvStack(p0: Int, p1: Int): ItemStack = invWrapper.takeInvStack(p0, p1)
-
-    override fun isInvEmpty(): Boolean = invWrapper.isInvEmpty
-
-    override fun isValidInvStack(int_1: Int, itemStack_1: ItemStack?): Boolean =
-        invWrapper.isValidInvStack(int_1, itemStack_1)
-
-    override val maxUpgrades: Int = MAX_UPGRADES
-
-    constructor(tag: CompoundTag) : this() {
-        fromTag(tag)
+    override fun popUpgrade(): Upgrade {
+        val result = appliedUpgrades.last()
+        markDirty()
+        appliedUpgrades = appliedUpgrades.dropLast(1)
+        return result
     }
 
     companion object {
         const val DEFAULT_CAPACITY_STACKS = 32
-        const val MAX_UPGRADES = 6
+        const val DEFAULT_MAX_UPGRADES = 6
 
         const val STORED_ITEM_TAG = "StoredItem"
         const val AMOUNT_STORED_TAG = "AmountStored"
         const val CLEAR_WHEN_EMPTY_TAG = "ClearWhenEmpty"
         const val UPGRADE_TAG = "Upgrades"
+        const val MAX_UPGRADES_TAG = "MaxUpgrades"
 
-        const val RIGHT_CLICK_PERIOD_MS = 500
+        const val STORED_BLOCK_ENTITY_TAG = "StoredBlockEntity"
 
         val TYPE: BlockEntityType<ItemBarrelBlockEntity> =
             BlockEntityType.Builder.create(Supplier { ItemBarrelBlockEntity() }, ItemBarrelBlock).build(null)
 
-        const val STORED_BLOCK_ENTITY_TAG = "StoredBlockEntity"
-
-        fun loadFromStack(stack: ItemStack): ItemBarrelBlockEntity =
-            ItemBarrelBlockEntity(stack.getOrCreateSubTag(STORED_BLOCK_ENTITY_TAG))
+        fun fromStack(stack: ItemStack): ItemBarrelBlockEntity {
+            val barrel = ItemBarrelBlockEntity()
+            barrel.fromClientTag(stack.getOrCreateSubTag(STORED_BLOCK_ENTITY_TAG))
+            return barrel
+        }
     }
 
-    val backingStorage get() = storage
+    val invAttribute = FixedMassItemInv(storage, false)
+    private val invWrapper = UnrestrictedInventoryFixedWrapper(invAttribute)
 
-    private var recentUsers: Map<UUID, Long> = mapOf()
-
-    fun onPunched(player: PlayerEntity) {
-        if (world?.isClient != false || storage.isEmpty) {
-            return
-        }
-
-        if (player.isSneaking) {
-            storage.take(storage.currentInstance.maxCount.toLong())[0].giveTo(player)
-        } else {
-            storage.take(1)[0].giveTo(player)
-        }
-
-        markDirty()
-        showContents(player)
-    }
-
-    fun onActivated(player: PlayerEntity) {
-        if (world?.isClient != false) {
-            return
-        }
-
-        recentUsers = recentUsers.filterValues { t -> System.currentTimeMillis() - t <= RIGHT_CLICK_PERIOD_MS }
-
-        if (player.isSneaking) {
-            if (storage.clearWhenEmpty) {
-                storage.retainInstanceWhenEmpty()
-                player.addChatMessage(TranslatableText("ui.stockpile.barrel.just_locked"), true)
-                player.playSound(SoundEvents.BLOCK_METAL_PRESSURE_PLATE_CLICK_ON, SoundCategory.BLOCKS, 0.1f, 0.9f)
-            } else {
-                storage.clearInstanceWhenEmpty()
-                player.addChatMessage(TranslatableText("ui.stockpile.barrel.just_unlocked"), true)
-                player.playSound(SoundEvents.BLOCK_METAL_PRESSURE_PLATE_CLICK_OFF, SoundCategory.BLOCKS, 0.1f, 0.9f)
-            }
-
-            markDirty()
-            return
-        }
-
-        recentUsers = if (player.uuid in recentUsers && storage.instanceIsSet) {
-            player.inventory.main.replaceAll { s -> storage.offer(s) }
-            player.inventory.offHand.replaceAll { s -> storage.offer(s) }
-            player.inventory.markDirty()
-
-            recentUsers - player.uuid
-        } else {
-            val activeStack = player.getStackInHand(Hand.MAIN_HAND)
-            val resultingStack = storage.offer(activeStack)
-
-            player.setStackInHand(Hand.MAIN_HAND, resultingStack)
-            player.inventory.markDirty()
-
-            recentUsers + (player.uuid to System.currentTimeMillis())
-        }
-
-        markDirty()
-        showContents(player)
-    }
-
-    private fun showContents(player: PlayerEntity) {
-        player.addChatMessage(getContentDescription(), true)
+    init {
+        invAttribute.addListener({ _, _, _, _ -> markDirty() }, { })
     }
 
     override fun markDirty() {
+        if (clearWhenEmpty && storage.isEmpty) {
+            storage.contents = ItemStackQuantizer.NONE
+        }
+
+        handleUpgradeChanges()
         super.markDirty()
-        invWrapper.markDirty()
 
         world?.apply {
             updateListeners(pos, getBlockState(pos), getBlockState(pos), 3)
         }
     }
 
-    override fun toTag(tag: CompoundTag?): CompoundTag {
-        return toTagWithoutWorldInfo(super.toTag(tag))
+    // TODO: Delegate to observable?
+    private fun handleUpgradeChanges() {
+        invAttribute.voidExtraItems = appliedUpgrades.filterIsInstance<TrashUpgrade>().any()
+
+        // TODO: Lame
+        (storage as MassItemStackStorage).maxStacks = appliedUpgrades.filterIsInstance<ItemBarrelUpgrade>()
+            .fold(DEFAULT_CAPACITY_STACKS) { acc, upgrade -> upgrade.upgradeMaxStacks(acc) }
     }
 
-    private fun toTagWithoutWorldInfo(tag: CompoundTag): CompoundTag {
-        tag.put(STORED_ITEM_TAG, storage.currentInstance.toTag(CompoundTag()))
-        tag.putLong(AMOUNT_STORED_TAG, storage.amountStored)
-        tag.putBoolean(CLEAR_WHEN_EMPTY_TAG, storage.clearWhenEmpty)
+    override fun giveToPlayer(player: PlayerEntity, amount: BarrelTransactionAmount) {
+        val removedItems = when (amount) {
+            BarrelTransactionAmount.ONE -> storage.contents.withAmount(1)
+            BarrelTransactionAmount.MANY -> storage.contents.reference.oneStackToQuantizer()
+            BarrelTransactionAmount.ALL -> TODO()
+        }
 
-        val upgradeTags = ListTag()
-        appliedUpgrades.forEach { u -> upgradeTags.add(UpgradeRegistry.writeUpgrade(u)) }
+        if (!removedItems.isEmpty) {
+            player.playSound(SoundEvents.ENTITY_ITEM_PICKUP, SoundCategory.BLOCKS, 0.5f, 0.8f)
+        }
 
-        tag.put(UPGRADE_TAG, upgradeTags)
+        storage.removeAtMost(removedItems).toObjects().forEach { it.giveTo(player) }
+        markDirty()
+    }
 
-        return tag
+    override fun takeFromPlayer(player: PlayerEntity, amount: BarrelTransactionAmount) {
+        var itemsTaken = false
+
+        when (amount) {
+            BarrelTransactionAmount.ONE -> TODO()
+
+            BarrelTransactionAmount.MANY -> {
+                if (!player.mainHandStack.isEmpty) {
+                    val result = storage.addAtMost(player.mainHandStack.toQuantizer()).firstStack()
+                    itemsTaken = player.mainHandStack.count != result.count
+
+                    player.setStackInHand(
+                        Hand.MAIN_HAND,
+                        result
+                    )
+                }
+            }
+
+            BarrelTransactionAmount.ALL -> {
+                player.inventory.main.replaceAll { storage.addAtMost(it.toQuantizer()).firstStack() }
+                itemsTaken = true
+            }
+        }
+
+        if (itemsTaken) {
+            player.playSound(SoundEvents.ENTITY_ITEM_PICKUP, SoundCategory.BLOCKS, 0.5f, 0.65f)
+        }
+
+        player.inventory.markDirty()
+        markDirty()
+    }
+
+    override fun changeModes(player: PlayerEntity) {
+        clearWhenEmpty = !clearWhenEmpty
+
+        if (clearWhenEmpty) {
+            player.addChatMessage(TranslatableText("ui.stockpile.barrel.just_unlocked"), true)
+            player.playSound(SoundEvents.BLOCK_LEVER_CLICK, SoundCategory.BLOCKS, 0.5f, 1.1f)
+        } else {
+            player.addChatMessage(TranslatableText("ui.stockpile.barrel.just_locked"), true)
+            player.playSound(SoundEvents.BLOCK_LEVER_CLICK, SoundCategory.BLOCKS, 0.5f, 0.9f)
+        }
+
+        markDirty()
+    }
+
+    override fun toTag(tag: CompoundTag?): CompoundTag {
+        requireNotNull(tag)
+        return toClientTag(super.toTag(tag))
     }
 
     override fun fromTag(tag: CompoundTag?) {
+        requireNotNull(tag)
+
         super.fromTag(tag)
-
-        if (tag == null) {
-            Stockpile.LOGGER.warn("an item barrel at $pos is missing data and will be reset to defaults")
-            return
-        }
-
-        fromTagWithoutWorldInfo(tag)
+        fromClientTag(tag)
     }
 
-    private fun fromTagWithoutWorldInfo(tag: CompoundTag) {
-        val storedItem = ItemStack.fromTag(tag.getCompound(STORED_ITEM_TAG))
-        val amountStored = if (storedItem.isEmpty) 0L else tag.getLong(AMOUNT_STORED_TAG)
-        val clearWhenEmpty = tag.getBoolean(CLEAR_WHEN_EMPTY_TAG)
-
-        val upgradeTags = tag.getList(UPGRADE_TAG, NbtType.COMPOUND).take(MAX_UPGRADES)
-
-        appliedUpgrades =
-            upgradeTags.mapNotNull { t -> (t as? CompoundTag)?.let { c -> UpgradeRegistry.readUpgrade(c) as? ItemBarrelUpgrade } }
-
-        val capacityStacks = appliedUpgrades.fold(DEFAULT_CAPACITY_STACKS) { i, u -> u.upgradeMaxStacks(i) }
-
-        storage = MassItemStorage(
-            capacityStacks,
-            min(amountStored, capacityStacks.toLong() * storedItem.maxCount),
-            storedItem,
-            clearWhenEmpty
-        )
-
-        invAttribute.storage = storage
-        invAttribute.voidExtraItems = appliedUpgrades.filterIsInstance<TrashUpgrade>().any()
+    override fun toClientTag(tag: CompoundTag?): CompoundTag = requireNotNull(tag).apply {
+        putString(STORED_ITEM_TAG, Registry.ITEM.getId(storage.contents.reference.item).toString())
+        putLong(AMOUNT_STORED_TAG, storage.contents.amount)
+        putBoolean(CLEAR_WHEN_EMPTY_TAG, clearWhenEmpty)
+        put(UPGRADE_TAG, appliedUpgrades.mapNotNull { u -> UpgradeRegistry.writeUpgrade(u) }.toCollection(ListTag()))
     }
 
-    override fun toClientTag(tag: CompoundTag?): CompoundTag = toTag(tag ?: CompoundTag())
-
-    override fun fromClientTag(tag: CompoundTag?) = fromTag(tag ?: CompoundTag())
-
-    override fun isUpgradeTypeAllowed(u: Upgrade): Boolean = u is ItemBarrelUpgrade
-
-    override fun pushUpgrade(u: Upgrade) {
-        if (u !is ItemBarrelUpgrade) {
-            Stockpile.LOGGER.warn("attempted to apply an invalid upgrade (type ${u.id}) to an item barrel")
-            Stockpile.LOGGER.debug("this is a bug! was isUpgradeTypeAllowed checked?")
-            return
-        }
-
-        appliedUpgrades += u
-
-        fromTagWithoutWorldInfo(toTagWithoutWorldInfo(CompoundTag()))
-        markDirty()
-    }
-
-    override fun popUpgrade(): Upgrade {
-        val result = appliedUpgrades.last()
-        appliedUpgrades = appliedUpgrades.dropLast(1)
-
-        fromTagWithoutWorldInfo(toTagWithoutWorldInfo(CompoundTag())) // TODO: what?
-        markDirty()
-
-        return result
-    }
-
-    fun toStack(stack: ItemStack) {
-        toTagWithoutWorldInfo(stack.getOrCreateSubTag(STORED_BLOCK_ENTITY_TAG))
-    }
-
-    fun fromStack(stack: ItemStack) {
-        fromTagWithoutWorldInfo(stack.getOrCreateSubTag(STORED_BLOCK_ENTITY_TAG))
-    }
-
-    fun getContentDescription(): Text {
-        val f = NumberFormat.getInstance()
-
-        return if (storage.isEmpty) {
-            TranslatableText("ui.stockpile.barrel.contents_empty", f.format(storage.maxStacks))
+    override fun fromClientTag(tag: CompoundTag?) = requireNotNull(tag).run {
+        // Upgrades
+        maxUpgrades = if (containsKey(MAX_UPGRADES_TAG)) {
+            getInt(MAX_UPGRADES_TAG)
         } else {
-            TranslatableText(
-                "ui.stockpile.barrel.contents",
-                f.format(storage.amountStored),
-                f.format(storage.capacity),
-                storage.currentInstance.name.asFormattedString(),
-                f.format(storage.amountStored / storage.currentInstance.maxCount),
-                f.format(storage.maxStacks)
-            )
+            DEFAULT_MAX_UPGRADES
+        }
+
+        appliedUpgrades = getList(UPGRADE_TAG, NbtType.COMPOUND)
+            .take(maxUpgrades)
+            .mapNotNull { t -> UpgradeRegistry.readUpgrade(t as? CompoundTag ?: return@mapNotNull null) }
+
+        // State
+        handleUpgradeChanges()
+        clearWhenEmpty = getBoolean(CLEAR_WHEN_EMPTY_TAG)
+
+        // Contents
+        val itemName = getString(STORED_ITEM_TAG)
+        val itemId = Identifier.tryParse(itemName) ?: Registry.ITEM.defaultId
+
+        if (itemName.isNullOrBlank() || itemId == Registry.ITEM.defaultId || !Registry.ITEM.containsId(itemId)) {
+            storage.contents = ItemStackQuantizer.NONE
+        } else {
+            val itemAmount = min(max(0L, getLong(AMOUNT_STORED_TAG)), storage.capacity)
+            storage.contents = ItemStackQuantizer(ItemStack(Registry.ITEM[itemId], 1), itemAmount)
         }
     }
+
+    // Delegation of Inventory to invAttribute. As far as I know we can't use Kotlin's implementation by delegation
+    // because the implementation can change.
+
+    override fun getInvStack(slot: Int): ItemStack = invWrapper.getInvStack(slot)
+
+    override fun clear() = invWrapper.clear()
+
+    override fun setInvStack(slot: Int, stack: ItemStack?) = invWrapper.setInvStack(slot, stack)
+
+    override fun removeInvStack(slot: Int): ItemStack = invWrapper.removeInvStack(slot)
+
+    override fun canPlayerUseInv(player: PlayerEntity?): Boolean = invWrapper.canPlayerUseInv(player)
+
+    override fun getInvSize(): Int = invWrapper.invSize
+
+    override fun takeInvStack(slot: Int, amount: Int): ItemStack = invWrapper.takeInvStack(slot, amount)
+
+    override fun isInvEmpty(): Boolean = invWrapper.isInvEmpty
+
+    override fun isValidInvStack(slot: Int, stack: ItemStack?): Boolean = invWrapper.isValidInvStack(slot, stack)
 }
+