+++ src/main/kotlin/me/branchpanic/mods/stockpile/content/upgrade/CapacityUpgrade.kt
@@ -6,6 +6,7 @@ import me.branchpanic.mods.stockpile.api.upgrade.UpgradeContainer
 import me.branchpanic.mods.stockpile.api.upgrade.UpgradeType
 import me.branchpanic.mods.stockpile.api.upgrade.barrel.ItemBarrelUpgrade
 import me.branchpanic.mods.stockpile.content.blockentity.ItemBarrelBlockEntity
+import me.branchpanic.mods.stockpile.impl.storage.MassItemStackStorage
 import net.minecraft.item.ItemStack
 import net.minecraft.nbt.CompoundTag
 import net.minecraft.text.Text
@@ -30,11 +31,12 @@ class CapacityUpgrade(private val amount: Int) : ItemBarrelUpgrade {
     override fun canSafelyBeRemovedFrom(context: UpgradeContainer): Boolean {
         val barrel = context as? ItemBarrelBlockEntity ?: return false
 
-        return barrel.backingStorage.amountStored <=
-                (barrel.backingStorage.maxStacks - amount) * barrel.backingStorage.storedStack.maxCount
+        return barrel.storage.contents.amount <=
+                ((barrel.storage as MassItemStackStorage).maxStacks - amount) *
+                barrel.storage.contents.reference.maxCount
     }
 
-    override fun getCorrespondingStack(): ItemStack = when (amount) {
+    override fun toStack(): ItemStack = when (amount) {
         32 -> ItemStack(Stockpile.ITEMS[id("capacity_upgrade")])
         64 -> ItemStack(Stockpile.ITEMS[id("double_capacity_upgrade")])
         else -> ItemStack.EMPTY