+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).  
 
+## [Unreleased]
+
+### Changed
+- Refactored barrel/storage APIs
+
+### Fixed
+- Item barrels not stacking after being crafted (now stack up to 8)
+
 ## [1.0.9] - 2019-07-21
 
 ### Added